business_trips_assistant/requirements.txt
aiogram==2.9
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.1
asyncpg==0.25.0
attrs==21.2.0
Babel==2.9.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==36.0.0
Django==3.2.9
django-cors-headers==3.10.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
frozenlist==1.2.0
h11==0.12.0
idna==3.3
lxml==4.6.4
multidict==5.2.0
numpy==1.21.4
opencv-python==4.5.4.60
outcome==1.1.0
Pillow==8.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
python-dotenv==0.19.2
pytz==2021.3
pyzbar==0.1.8
reportlab==3.6.3
requests==2.26.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.0
urllib3==1.26.7
wsproto==1.0.0
yarl==1.7.2
